% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcq.R
\docType{package}
\name{mcq}
\alias{mcq}
\title{mcq: A package for evaluating multiple choice test results}
\description{
The mcq package provides three functions:
score_a_questions() for 'A' questions, score_k_questions() for 'Kprime'
and score_exams(), which calls both of the above and joins their outputs.
}
\section{mcq functions}{

score_a_questions()
score_k_questions()
score_exams()
}

